Move away from Kerbal SpacePort whenever possible.
Kerbal SpacePort shuts down May 30, so it is the least-desirable download
location.

The following mods remain on SpacePort:
  - Kerbal Attachment System
  - Mk2 Cockpit Interior
  - Remote Tech 2
  - Vanguard Technologies

Kerbal Attachment System has a weird license so I don't know if it'd be legal to
reproduce that one, but the other ones I could always host myself (or with
GitHub or whatever).