Add support for a download cache.
This is helpful when I want to avoid having to hit external downloads, because
I'm running into some issues with sites throttling me.

This is a development feature, not meant to be used by end users. Most users do
not download packages repeatedly, but those who are testing their downloads
will.